prestashop: workaround bug when updating last product of a category
Prestashop's handling of in-category ordering is a non-sense, at least
on the WebService part. Bug BOOM-5826 makes the WS reject a value it
sent itself…
So currently the best we can do is to pass the update by putting the
position_in_category to zero as we cannot do anything else (except
fetching the whole category and adjust position if needed, which is a
no-go). The consequence it that each updated product is put at the first
place of its category.
Fixes #25