prestashop: rework discount handling
Modify the way discounts are handled to ensure computed total always
matches prestashop one. That means, we do not rely on any discount field
coming from prestashop, we simply compute the delta between total per
line and order total and add an adjustment line if needed. It involves a
new configuration option in the module in order to select which product
is used for discount.